rename rubicon_ethereum to rubicon
